修改地址

+++ iseeu/ClassificationTableView.m
@@ -427,7 +427,7 @@
     
     RightViewModel *rightViewModel = (RightViewModel *)self.fakeData[indexPath.row];
     
-    [homeDelegate pushClassificationViewController:self.tag withSid:[rightViewModel.id intValue]];
+    [homeDelegate pushClassificationViewController:(int)self.tag withSid:[rightViewModel.id intValue]];
     
 }
